refactor: remove control-plane core modules from federated-catalog
