+++ core/common/connector-core/src/main/java/org/eclipse/edc/connector/core/profile/DataspaceProfileContextRegistryImpl.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2025 Cofinity-X
+ *  Copyright (c) 2026 Think-it GmbH
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,11 +8,11 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Cofinity-X - initial API and implementation
+ *       Think-it GmbH - initial API and implementation
  *
  */
 
-package org.eclipse.edc.connector.controlplane.profile;
+package org.eclipse.edc.connector.core.profile;
 
 
 import org.eclipse.edc.protocol.spi.DataspaceProfileContext;
