+++ core/control-plane/control-plane-core/build.gradle.kts
@@ -23,9 +23,9 @@ dependencies {
 
     implementation(project(":core:control-plane:lib:control-plane-lib"))
     implementation(project(":core:common:boot"))
-    implementation(project(":core:common::participant-context-core"))
-    implementation(project(":core:common::participant-context-connector-core"))
-    implementation(project(":core:common::participant-context-config-core"))
+    implementation(project(":core:common:participant-context-core"))
+    implementation(project(":core:common:participant-context-connector-core"))
+    implementation(project(":core:common:participant-context-config-core"))
     implementation(project(":core:control-plane:control-plane-catalog"))
     implementation(project(":core:control-plane:control-plane-contract"))
     implementation(project(":core:control-plane:control-plane-transfer"))
