+++ core/control-plane/control-plane-core/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -13,5 +13,3 @@
 #
 
 org.eclipse.edc.connector.controlplane.ControlPlaneDefaultServicesExtension
-org.eclipse.edc.connector.controlplane.discovery.DiscoveryDefaultServicesExtension
-org.eclipse.edc.connector.controlplane.discovery.DiscoveryServicesExtension
