+++ extensions/common/events/events-cloud-http/build.gradle.kts
@@ -24,7 +24,6 @@ dependencies {
     implementation(libs.failsafe.core)
 
     testImplementation(testFixtures(project(":core:common:junit")))
-    testImplementation(project(":core:common:connector-core"))
     testImplementation(project(":core:common:runtime-core"))
 
     testImplementation(libs.wiremock)
